Removed warnings.

+++ Sonic/Sonic.Server/entities/Server.cpp
@@ -135,7 +135,7 @@ int Server::getAvailableIndex()
 
 int Server::getDisconnectedIndex()
 {
-	for (size_t i = 0; i < this->serverConfig->getMaxAllowedClients(); i++)
+	for (int i = 0; i < this->serverConfig->getMaxAllowedClients(); i++)
 	{
 		if (clients.count(i)) 
 		{
